+++ src/test/java/persistence/entity/SingleTableEntityPersisterTest.java
@@ -3,6 +3,8 @@ package persistence.entity;
 import jdbc.RowMapper;
 import org.junit.jupiter.api.*;
 import persistence.JdbcServerDmlQueryTestSupport;
+import persistence.entity.persister.EntityPersister;
+import persistence.entity.persister.SingleTableEntityPersister;
 import persistence.sql.ddl.PersonV3;
 import persistence.sql.dialect.Dialect;
 import persistence.sql.dialect.H2Dialect;
